fixes with env and add modal window to frontside

+++ web/app/Lib/Handlers/AppUninstalled.php
@@ -14,9 +14,19 @@ class AppUninstalled implements Handler
     public function handle(string $topic, string $shop, array $body): void
     {
         Log::debug("App was uninstalled from $shop - removing all sessions and scripts");
-        $file = ScriptTag::where('shop', $shop)->value('script_file');
-        unlink($file);
-        ScriptTag::where('shop', $shop)->delete();
+        self::deleteScriptTag($shop);
         Session::where('shop', $shop)->delete();
+    }
+
+    public static function deleteScriptTag(string $shop)
+    {
+        if (ScriptTag::where('shop', $shop)->first()) {
+            $file = ScriptTag::where('shop', $shop)->value('script_file');
+            unlink($file);
+            ScriptTag::where('shop', $shop)->delete();
+            Log::debug("Script tag removed for $shop");
+        } else {
+            Log::debug("Script tag for $shop not found. Nothing to delete");
+        }
     }
 }
